created new dict_helpers.py file
