+++ usaspending_api/common/cache.py
@@ -4,7 +4,7 @@ import json
 from rest_framework_extensions.key_constructor import bits
 from rest_framework_extensions.key_constructor.constructors import DefaultKeyConstructor
 
-from usaspending_api.common.helpers.generic_helper import order_nested_object
+from usaspending_api.common.helpers.dict_helpers import order_nested_object
 
 
 class PathKeyBit(bits.QueryParamsKeyBit):
