+++ usaspending_api/common/helpers/generic_helper.py
@@ -4,7 +4,6 @@ import subprocess
 import time
 
 from calendar import monthrange, isleap
-from collections import OrderedDict
 from datetime import datetime as dt
 from django.conf import settings
 from django.db import DEFAULT_DB_ALIAS, connection
@@ -42,12 +41,6 @@ def read_text_file(filepath):
     with open(filepath, "r") as plaintext_file:
         file_content_str = plaintext_file.read()
     return file_content_str
-
-
-def upper_case_dict_values(input_dict):
-    for key in input_dict:
-        if isinstance(input_dict[key], str):
-            input_dict[key] = input_dict[key].upper()
 
 
 def validate_date(date):
@@ -217,32 +210,6 @@ def generate_raw_quoted_query(queryset):
             str_fix_param = param
         str_fix_params.append(str_fix_param)
     return sql % tuple(str_fix_params)
-
-
-def order_nested_object(nested_object):
-    """
-    Simply recursively order the item. To be used for standardizing objects for JSON dumps
-    """
-    if isinstance(nested_object, list):
-        if len(nested_object) > 0 and isinstance(nested_object[0], dict):
-            # Lists of dicts aren't handled by python's sorted(), so we handle sorting manually
-            sorted_subitems = []
-            sort_dict = {}
-            # Create a hash using keys & values
-            for subitem in nested_object:
-                hash_list = ['{}{}'.format(key, subitem[key]) for key in sorted(list(subitem.keys()))]
-                hash_str = '_'.join(str(hash_list))
-                sort_dict[hash_str] = order_nested_object(subitem)
-            # Sort by the new hash
-            for sorted_hash in sorted(list(sort_dict.keys())):
-                sorted_subitems.append(sort_dict[sorted_hash])
-            return sorted_subitems
-        else:
-            return sorted([order_nested_object(subitem) for subitem in nested_object])
-    elif isinstance(nested_object, dict):
-        return OrderedDict([(key, order_nested_object(nested_object[key])) for key in sorted(nested_object.keys())])
-    else:
-        return nested_object
 
 
 def generate_matviews():
