+++ usaspending_api/download/management/commands/populate_monthly_files.py
@@ -6,7 +6,8 @@ import re
 
 from django.conf import settings
 from django.core.management.base import BaseCommand
-from usaspending_api.common.helpers.generic_helper import generate_fiscal_year, order_nested_object
+from usaspending_api.common.helpers.dict_helpers import order_nested_object
+from usaspending_api.common.helpers.generic_helper import generate_fiscal_year
 from usaspending_api.common.sqs_helpers import get_sqs_queue_resource
 from usaspending_api.download.filestreaming import csv_generation
 from usaspending_api.download.helpers import multipart_upload, pull_modified_agencies_cgacs
